COMMERCE-6553 Source formatting
