+++ modules/apps/commerce/commerce-product-api/src/main/java/com/liferay/commerce/product/constants/CPConstants.java
@@ -45,13 +45,13 @@ public class CPConstants {
 	public static final String RESOURCE_NAME_CHANNEL =
 		"com.liferay.commerce.channel";
 
+	public static final String RESOURCE_NAME_CP_DISPLAY_LAYOUT =
+		"com.liferay.commerce.product.display.layout";
+
 	public static final String RESOURCE_NAME_PRODUCT =
 		"com.liferay.commerce.product";
 
 	public static final String RESOURCE_NAME_TAX = "com.liferay.commerce.tax";
-
-	public static final String RESOURCE_NAME_CP_DISPLAY_LAYOUT =
-		"com.liferay.commerce.product.display.layout";
 
 	public static final String SERVICE_NAME_CP_FRIENDLY_URL =
 		"com.liferay.commerce.product.friendly.url";
