+++ modules/apps/commerce/commerce-product-service/src/main/java/com/liferay/commerce/product/internal/settings/definition/CPDisplayLayoutConfigurationBeanDeclaration.java
@@ -12,29 +12,23 @@
  * details.
  */
 
-package com.liferay.commerce.product.internal.configuration;
+package com.liferay.commerce.product.internal.settings.definition;
 
 import com.liferay.commerce.product.configuration.CPDisplayLayoutConfiguration;
-import com.liferay.commerce.product.configuration.CPFriendlyURLConfiguration;
-import com.liferay.commerce.product.constants.CPConstants;
-import com.liferay.portal.kernel.settings.definition.ConfigurationPidMapping;
+import com.liferay.portal.kernel.settings.definition.ConfigurationBeanDeclaration;
+
 import org.osgi.service.component.annotations.Component;
 
 /**
  * @author Alec Sloan
  */
-@Component(enabled = false, service = ConfigurationPidMapping.class)
-public class CPDisplayLayoutConfigurationPidMapping
-	implements ConfigurationPidMapping {
+@Component(enabled = false, service = ConfigurationBeanDeclaration.class)
+public class CPDisplayLayoutConfigurationBeanDeclaration
+	implements ConfigurationBeanDeclaration {
 
 	@Override
 	public Class<?> getConfigurationBeanClass() {
 		return CPDisplayLayoutConfiguration.class;
-	}
-
-	@Override
-	public String getConfigurationPid() {
-		return CPConstants.RESOURCE_NAME_CP_DISPLAY_LAYOUT;
 	}
 
 }
